[Issue #183] Move getValidationConstraints() to exported module
