+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Additions
 - [Issue #187] Add HttpClient::request for proxy
+- [Issue #183] Move getValidationConstraints() to exported module
 
 ### Fixes
 - npm update
+- Cleaned up query-tools
 
 ## [3.0.3] Jul-23-2019
